From 2b6ad08fb3db28bd4596873db0049b265ccfa688 Mon Sep 17 00:00:00 2001 From: Sylvain Munaut <tnt@246tNt.com> Date: Thu, 15 May 2014 11:16:19 +0200 Subject: [PATCH] main: Request Qt to setup XInitThreads
This allows multi-threaded GUI access on X11 platforms. Windows and
OSX platforms allow multi-thread by default, no need for special
setup. This option is a no-op on those platform, so no harm in calling
it for all platforms.

This is only available on Qt >= 4.8, so guard it with a version
check. This also means that any feature actually requiring multithreaded
GUI access need to be disabled for Qt < 4.8 !!!

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>